+++ fastiron/src/parameters.rs
@@ -8,7 +8,7 @@ use rustc_hash::FxHashMap;
 
 use crate::{
     constants::CustomFloat,
-    utils::io_utils::{parse_input_file, Cli, InputError},
+    utils::input::{parse_input_file, Cli, InputError},
 };
 
 /// Alias for a `<String, String>` [`FxHashMap`]. See here for detailed
