+++ fastiron/src/utils/mod.rs
@@ -3,7 +3,7 @@
 pub mod comm_object;
 pub mod coral_benchmark_correctness;
 pub mod decomposition_object;
-pub mod io_utils;
+pub mod input;
 pub mod mc_fast_timer;
 pub mod mc_processor_info;
 pub mod mc_rng_state;
